pyproject.toml
[project]
name = "trading-ml-pipeline"
version = "0.1.0"
description = "Local ML pipeline for evaluating stock trading strategies"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "polars>=1.0",
    "psycopg[binary]>=3.2",
    "sqlalchemy>=2.0",
    "torch>=2.4",
    "lightning>=2.4",
    "mlflow>=2.16",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "rich>=14.3.3",
    "scikit-learn>=1.5",
    "scipy>=1.14",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.6",
    "mypy>=1.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
